@@ -2,10 +2,10 @@ name: nyanya_rocket
 description: Save cute mice from dangerous cats!
 homepage: https://github.com/CaramelDunes/nyanya_rocket
 
-version: 1.1.0+45
+version: 2.0.0+50
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.5.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -18,30 +18,31 @@ dependencies:
   flutter_localizations:
     sdk: flutter
 
-  intl: ^0.16.0
+  intl: ^0.16.1
 
-  firebase_core: ^0.4.3+2
-  cloud_firestore: ^0.13.0+1
-  firebase_auth: ^0.15.3+1
-  cloud_functions: ^0.4.1+6
+  firebase_core: ^0.5.0
+  cloud_firestore: ^0.14.0+2
+  firebase_auth: ^0.18.0+1
+  cloud_functions: ^0.6.0
 
-  flare_flutter: ^1.8.3
+  flare_flutter: ^2.0.1
 
-  font_awesome_flutter: ^8.5.0
+  font_awesome_flutter: ^8.7.0
 
-  shared_preferences: ^0.5.1+2
+  shared_preferences: ^0.5.10
 
-  path_provider: ^1.5.1
+  path_provider: ^1.6.14
 
   uuid: ^2.0.4
 
-  connectivity: ^0.4.6+2
+  connectivity: ^0.4.9+2
 
-  url_launcher: ^5.4.1
+  url_launcher: ^5.4.2
 
-  share: ^0.6.3+5
+  share: ^0.6.4+5
 
-  provider: ^2.0.1
+  provider: ^4.3.2+1
+  characters: ^1.0.0
 
 dev_dependencies:
   flutter_test:
@@ -49,7 +50,7 @@ dev_dependencies:
 
   flutter_launcher_icons: ^0.7.4
 
-  intl_translation: ^0.17.4
+  intl_translation: ^0.17.9
 
 flutter_icons:
   image_path: "assets/graphics/app_icon.png"
