@@ -3,7 +3,7 @@ import 'package:nyanya_rocket/blocs/local_game_controller.dart';
 import 'package:nyanya_rocket_base/nyanya_rocket_base.dart';
 
 class EditorGameController extends LocalGameController {
-  EditorGameController({@required Game game}) : super(game) {
+  EditorGameController({@required GameState game, @required GameSimulator gameSimulator}) : super(game, gameSimulator) {
     running = false;
   }
 
@@ -20,18 +20,28 @@ class EditorGameController extends LocalGameController {
   bool toggleEntity(int x, int y, EntityType type, Direction direction) {
     if (!running) {
       if (game.board.tiles[x][y] is Empty || game.board.tiles[x][y] is Arrow) {
-        Iterable<Entity> there = game.entities.where((Entity entity) =>
+        Iterable<Entity> miceThere = game.mice.where((Entity entity) =>
             entity.position.x == x && entity.position.y == y);
 
-        if (there.isEmpty) {
-          game.entities.add(Entity.fromEntityType(
-              type, BoardPosition.centered(x, y, direction)));
+        Iterable<Entity> catsThere = game.cats.where((Entity entity) =>
+            entity.position.x == x && entity.position.y == y);
+
+        if (miceThere.isEmpty && catsThere.isEmpty) {
+          if (type == EntityType.Cat) {
+            game.cats.add(Entity.fromEntityType(
+                type, BoardPosition.centered(x, y, direction)));
+          } else {
+            game.mice.add(Entity.fromEntityType(
+                type, BoardPosition.centered(x, y, direction)));
+          }
 
           updateGame();
           return true;
-        } else if ((there.first is Mouse && type == EntityType.Mouse) ||
-            (there.first is Cat && type == EntityType.Cat)) {
-          game.entities.remove(there.first);
+        } else if (miceThere.isNotEmpty && type == EntityType.Mouse) {
+          game.mice.remove(miceThere.first);
+          updateGame();
+        } else if (catsThere.isNotEmpty is Cat && type == EntityType.Cat) {
+          game.cats.remove(catsThere.first);
           updateGame();
         }
       }
