@@ -19,6 +19,18 @@ typedef String MessageIfAbsent(String messageStr, List<dynamic> args);
 class MessageLookup extends MessageLookupByLibrary {
   String get localeName => 'en';
 
+  static m0(percentCompleted) => "${percentCompleted}% completed";
+
+  static m1(minuteCount) => "${minuteCount} minutes";
+
+  static m2(playerCount) => "${playerCount} players";
+
+  static m3(queueLength) => "${Intl.plural(queueLength, zero: 'No player in queue', other: '${queueLength} players in queue.')}";
+
+  static m4(position, queueLength) => "Position in queue: ${position} / ${queueLength}";
+
+  static m5(localIp) => "This device\'s IP: ${localIp}";
+
   final messages = _notInlinedMessages(_notInlinedMessages);
   static _notInlinedMessages(_) => <String, Function> {
     "accept" : MessageLookupByLibrary.simpleMessage("Accept"),
@@ -27,7 +39,9 @@ class MessageLookup extends MessageLookupByLibrary {
     "arrowRemoveTutorialText" : MessageLookupByLibrary.simpleMessage("Arrows can be removed by tapping"),
     "arrowSwipeTutorialText" : MessageLookupByLibrary.simpleMessage("...or swiped"),
     "arrowTutorialText" : MessageLookupByLibrary.simpleMessage("Arrow"),
+    "awaitingForPlayersLabel" : MessageLookupByLibrary.simpleMessage("Awaiting for players..."),
     "back" : MessageLookupByLibrary.simpleMessage("Back"),
+    "boardLabel" : MessageLookupByLibrary.simpleMessage("Board"),
     "boardSelectionText" : MessageLookupByLibrary.simpleMessage("Tap to select"),
     "cancel" : MessageLookupByLibrary.simpleMessage("Cancel"),
     "catTutorialText" : MessageLookupByLibrary.simpleMessage("Cats are slower than mice.\nWhen a cat meets an arrow head-on, it will damage it or destroy it if already damaged."),
@@ -42,33 +56,41 @@ class MessageLookup extends MessageLookupByLibrary {
     "challengeType" : MessageLookupByLibrary.simpleMessage("Challenge"),
     "challengesTitle" : MessageLookupByLibrary.simpleMessage("Challenges"),
     "communityTab" : MessageLookupByLibrary.simpleMessage("Community"),
-    "completedLabel" : MessageLookupByLibrary.simpleMessage("% Completed"),
+    "completedPercentLabel" : m0,
     "confirmLabel" : MessageLookupByLibrary.simpleMessage("Confirm"),
     "connectedStatusLabel" : MessageLookupByLibrary.simpleMessage("Connected"),
+    "connectingToServerText" : MessageLookupByLibrary.simpleMessage("Connecting to server..."),
     "contributingTab" : MessageLookupByLibrary.simpleMessage("Contributing"),
     "contributingText" : MessageLookupByLibrary.simpleMessage("Help NyaNya Rocket!\n\nYou don\'t need any particular skills to contribute, you can:\n  • Report bugs and ask for new features\n  • Translate\n  • Improve graphical assets and animation\n  • Code\n\nFor more information, head over to "),
     "contributorsLabel" : MessageLookupByLibrary.simpleMessage("Contributors"),
     "createLabel" : MessageLookupByLibrary.simpleMessage("Create"),
     "darkModeLabel" : MessageLookupByLibrary.simpleMessage("Dark Mode"),
     "dateLabel" : MessageLookupByLibrary.simpleMessage("Date"),
+    "defeatLabel" : MessageLookupByLibrary.simpleMessage("Defeat..."),
     "deny" : MessageLookupByLibrary.simpleMessage("Deny"),
+    "deviceDuelLabel" : MessageLookupByLibrary.simpleMessage("Device duel"),
     "deviceTab" : MessageLookupByLibrary.simpleMessage("Device"),
     "discardDialogMessage" : MessageLookupByLibrary.simpleMessage("Are you sure you want to leave?\nAny unsaved changes will be discarded!"),
     "discardDialogTitle" : MessageLookupByLibrary.simpleMessage("Confirm leave"),
     "disconnectedStatusLabel" : MessageLookupByLibrary.simpleMessage("Disconnected"),
     "displayNameChangeSuccessText" : MessageLookupByLibrary.simpleMessage("Display name successfully changed!"),
-    "displayNameDialogTitle" : MessageLookupByLibrary.simpleMessage("Please enter your new display name"),
+    "displayNameDialogTitle" : MessageLookupByLibrary.simpleMessage("Choose your display name"),
     "displayNameFormatText" : MessageLookupByLibrary.simpleMessage("Between 2 and 24 characters (no space)."),
     "displayNameLabel" : MessageLookupByLibrary.simpleMessage("Display name"),
+    "duelLabel" : MessageLookupByLibrary.simpleMessage("Duel"),
+    "durationLabel" : MessageLookupByLibrary.simpleMessage("Duration"),
     "easyLabel" : MessageLookupByLibrary.simpleMessage("Easy"),
     "editTab" : MessageLookupByLibrary.simpleMessage("Edit"),
     "editorTitle" : MessageLookupByLibrary.simpleMessage("Editor"),
     "emptyListText" : MessageLookupByLibrary.simpleMessage("Nothing here..."),
     "enableAnimationsLabel" : MessageLookupByLibrary.simpleMessage("Enable Animations"),
     "entitiesTutorialLabel" : MessageLookupByLibrary.simpleMessage("Entities"),
+    "findPlayersLabel" : MessageLookupByLibrary.simpleMessage("Find players"),
     "firstTimeButtonLabel" : MessageLookupByLibrary.simpleMessage("Take me to it!"),
     "firstTimeText" : MessageLookupByLibrary.simpleMessage("First time here? Check the tutorial!"),
     "firstTimeWelcome" : MessageLookupByLibrary.simpleMessage("Welcome,"),
+    "fourPlayersLabel" : MessageLookupByLibrary.simpleMessage("4-players Battle"),
+    "friendDuelLabel" : MessageLookupByLibrary.simpleMessage("Friend Duel"),
     "generalLabel" : MessageLookupByLibrary.simpleMessage("General"),
     "getMiceObjectiveText" : MessageLookupByLibrary.simpleMessage("Lead all mice to the goal within 30 sec."),
     "hardLabel" : MessageLookupByLibrary.simpleMessage("Hard"),
@@ -79,13 +101,18 @@ class MessageLookup extends MessageLookupByLibrary {
     "invalidHostnameText" : MessageLookupByLibrary.simpleMessage("Please enter a valid hostname."),
     "invalidNameText" : MessageLookupByLibrary.simpleMessage("Please enter a valid name."),
     "invalidNicknameText" : MessageLookupByLibrary.simpleMessage("Please enter a valid nickname."),
-    "lanTab" : MessageLookupByLibrary.simpleMessage("LAN"),
+    "joinQueueLabel" : MessageLookupByLibrary.simpleMessage("Join"),
+    "joinRoomLabel" : MessageLookupByLibrary.simpleMessage("Join a room"),
+    "lanMultiplayerLabel" : MessageLookupByLibrary.simpleMessage("LAN multiplayer"),
     "languageLabel" : MessageLookupByLibrary.simpleMessage("Language"),
+    "leaderBoardLabel" : MessageLookupByLibrary.simpleMessage("Leaderboard"),
     "loadingLabel" : MessageLookupByLibrary.simpleMessage("Loading..."),
-    "loginPromptText" : MessageLookupByLibrary.simpleMessage("Please sign-in first! (Settings—Account Management)"),
+    "loginButtonLabel" : MessageLookupByLibrary.simpleMessage("Sign-in"),
+    "loginPromptText" : MessageLookupByLibrary.simpleMessage("Please sign-in first."),
     "loginStatusLabel" : MessageLookupByLibrary.simpleMessage("Login status"),
     "lunchTimeObjectiveText" : MessageLookupByLibrary.simpleMessage("Feed the cat!"),
     "matchmakingTab" : MessageLookupByLibrary.simpleMessage("Matchmaking"),
+    "minuteCountLabel" : m1,
     "mouseTutorialText" : MessageLookupByLibrary.simpleMessage("Mouse"),
     "movementTutorialText" : MessageLookupByLibrary.simpleMessage("\nAn entity walks straight until it encounters an arrow or a wall.\nFor some reasons, mice and cats prefer turning right when hitting a wall...\n"),
     "multiplayerTitle" : MessageLookupByLibrary.simpleMessage("Multiplayer"),
@@ -95,23 +122,33 @@ class MessageLookup extends MessageLookupByLibrary {
     "newsLabel" : MessageLookupByLibrary.simpleMessage("News"),
     "nextLevelLabel" : MessageLookupByLibrary.simpleMessage("Next level"),
     "nicknameLabel" : MessageLookupByLibrary.simpleMessage("Nickname"),
+    "nicknamesLabel" : MessageLookupByLibrary.simpleMessage("Nicknames"),
     "no" : MessageLookupByLibrary.simpleMessage("No"),
     "notImplementedText" : MessageLookupByLibrary.simpleMessage("Not (yet) available!\nWant it to be a priority? :-)"),
     "oneHundredMiceObjectiveText" : MessageLookupByLibrary.simpleMessage("Collect 100 mice in 30 sec."),
     "originalTab" : MessageLookupByLibrary.simpleMessage("Original"),
     "pitTutorialText" : MessageLookupByLibrary.simpleMessage("Black Hole (technically a square though...)"),
     "placementTutorialLabel" : MessageLookupByLibrary.simpleMessage("Arrow placement"),
+    "playAgainLabel" : MessageLookupByLibrary.simpleMessage("Play again"),
     "playLabel" : MessageLookupByLibrary.simpleMessage("Play"),
-    "playersLabel" : MessageLookupByLibrary.simpleMessage("players"),
+    "playerCountLabel" : m2,
+    "playersInQueueText" : m3,
     "popularityLabel" : MessageLookupByLibrary.simpleMessage("Popularity"),
+    "positionInQueueText" : m4,
     "privacyPolicyLabel" : MessageLookupByLibrary.simpleMessage("Privacy Policy"),
+    "privacyPolicySignUpText" : MessageLookupByLibrary.simpleMessage("By clicking Confirm, you agree to the "),
     "publishLabel" : MessageLookupByLibrary.simpleMessage("Publish"),
     "publishSuccessText" : MessageLookupByLibrary.simpleMessage("Puzzle published!"),
     "puzzleNotCompletedLocallyText" : MessageLookupByLibrary.simpleMessage("Puzzle not published as you didn\'t complete it"),
     "puzzleTutorialText" : MessageLookupByLibrary.simpleMessage("A Puzzle game is won when every mice has reached a rocket.\nYou have a restricted amount of arrows to achieve that goal.\n\nNote that you cannot place or remove an arrow after pressing the play button; you\'ll need to reset the level first.\n\nYou\'ll lose if:\n  • A cat enters a rocket before every mice is safe.\n  • A mouse gets eaten by a cat.\n  • A mouse falls into a black hole.\n\nYou\'ll be awarded a star if you complete a level without using all the available arrows.\n"),
     "puzzleType" : MessageLookupByLibrary.simpleMessage("Puzzle"),
     "puzzlesTitle" : MessageLookupByLibrary.simpleMessage("Puzzles"),
+    "queueRefreshErrorText" : MessageLookupByLibrary.simpleMessage("Error while refreshing queue info."),
+    "refreshLabel" : MessageLookupByLibrary.simpleMessage("Refresh"),
+    "regionLabel" : MessageLookupByLibrary.simpleMessage("Region"),
     "rocketTutorialText" : MessageLookupByLibrary.simpleMessage("Rocket"),
+    "roomCodeLabel" : MessageLookupByLibrary.simpleMessage("Room code"),
+    "roomCodeRetrievalErrorText" : MessageLookupByLibrary.simpleMessage("Could not retrieve room code."),
     "runAwayObjectiveText" : MessageLookupByLibrary.simpleMessage("Avoid the cat and get to the goal!"),
     "saveLabel" : MessageLookupByLibrary.simpleMessage("Save"),
     "settingsTitle" : MessageLookupByLibrary.simpleMessage("Settings"),
@@ -124,11 +161,16 @@ class MessageLookup extends MessageLookupByLibrary {
     "sortByLabel" : MessageLookupByLibrary.simpleMessage("Sort by"),
     "stageClearedText" : MessageLookupByLibrary.simpleMessage("Stage cleared!"),
     "tapToChangeDisplayNameLabel" : MessageLookupByLibrary.simpleMessage("Tap to change"),
+    "thisDeviceIpText" : m5,
     "tilesTutorialLabel" : MessageLookupByLibrary.simpleMessage("Tiles"),
     "tutorialTitle" : MessageLookupByLibrary.simpleMessage("How to play"),
     "unauthenticatedError" : MessageLookupByLibrary.simpleMessage("Error: Unauthenticated"),
     "veryHardLabel" : MessageLookupByLibrary.simpleMessage("Very Hard"),
+    "victoryLabel" : MessageLookupByLibrary.simpleMessage("Victory!"),
+    "waitingForPlayersText" : MessageLookupByLibrary.simpleMessage("Waiting for players..."),
     "whatsNewTab" : MessageLookupByLibrary.simpleMessage("What\'s New?"),
-    "yes" : MessageLookupByLibrary.simpleMessage("Yes")
+    "winnerLabel" : MessageLookupByLibrary.simpleMessage("Winner!"),
+    "yes" : MessageLookupByLibrary.simpleMessage("Yes"),
+    "yourRoomCodeLabel" : MessageLookupByLibrary.simpleMessage("Your room code is")
   };
 }
