@@ -9,6 +9,9 @@ import 'package:nyanya_rocket/widgets/default_drawer/default_drawer.dart';
 class Challenges extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
+    final bool displayIcons =
+        MediaQuery.of(context).orientation == Orientation.portrait;
+
     return DefaultTabController(
       length: 3,
       child: Scaffold(
@@ -17,15 +20,17 @@ class Challenges extends StatelessWidget {
             bottom: TabBar(
               tabs: [
                 Tab(
-                  icon: Icon(FontAwesomeIcons.stopwatch),
+                  icon:
+                      displayIcons ? FaIcon(FontAwesomeIcons.stopwatch) : null,
                   text: NyaNyaLocalizations.of(context).originalTab,
                 ),
                 Tab(
-                  icon: Icon(FontAwesomeIcons.globe),
+                  icon: displayIcons ? FaIcon(FontAwesomeIcons.globe) : null,
                   text: NyaNyaLocalizations.of(context).communityTab,
                 ),
                 Tab(
-                  icon: Icon(FontAwesomeIcons.mobileAlt),
+                  icon:
+                      displayIcons ? FaIcon(FontAwesomeIcons.mobileAlt) : null,
                   text: NyaNyaLocalizations.of(context).deviceTab,
                 ),
               ],
