@@ -10,9 +10,7 @@ import 'package:nyanya_rocket/widgets/success_overlay.dart';
 
 class CommunityChallenges extends StatefulWidget {
   @override
-  _CommunityChallengesState createState() {
-    return _CommunityChallengesState();
-  }
+  _CommunityChallengesState createState() => _CommunityChallengesState();
 }
 
 enum _Sorting { ByDate, ByPopularity, ByName }
@@ -28,7 +26,7 @@ class _CommunityChallengesState extends State<CommunityChallenges> {
   }
 
   Future<void> _refreshList() async {
-    QuerySnapshot snapshot = await Firestore.instance
+    QuerySnapshot snapshot = await FirebaseFirestore.instance
         .collection('challenges')
         .orderBy(
             _sorting == _Sorting.ByDate
@@ -36,18 +34,18 @@ class _CommunityChallengesState extends State<CommunityChallenges> {
                 : _sorting == _Sorting.ByPopularity ? 'likes' : 'name',
             descending: _sorting != _Sorting.ByName)
         .limit(50)
-        .getDocuments();
+        .get();
 
-    List<CommunityChallengeData> newChallenges = snapshot.documents
-        .map<CommunityChallengeData>((DocumentSnapshot snapshot) {
+    List<CommunityChallengeData> newChallenges =
+        snapshot.docs.map<CommunityChallengeData>((DocumentSnapshot snapshot) {
       return CommunityChallengeData(
-          uid: snapshot.documentID,
+          uid: snapshot.id,
           challengeData: ChallengeData.fromJson(
-              jsonDecode(snapshot.data['challenge_data'])),
-          likes: snapshot.data['likes'],
-          author: snapshot.data['author_name'],
-          name: snapshot.data['name'],
-          date: snapshot.data['date'].toDate());
+              jsonDecode(snapshot.get('challenge_data'))),
+          likes: snapshot.get('likes'),
+          author: snapshot.get('author_name'),
+          name: snapshot.get('name'),
+          date: snapshot.get('date').toDate());
     }).toList();
 
     if (mounted) {
@@ -69,7 +67,7 @@ class _CommunityChallengesState extends State<CommunityChallenges> {
             children: <Widget>[
               Text(
                 NyaNyaLocalizations.of(context).sortByLabel,
-                style: Theme.of(context).textTheme.subhead,
+                style: Theme.of(context).textTheme.subtitle1,
               ),
               VerticalDivider(),
               Expanded(
@@ -127,12 +125,12 @@ class _CommunityChallengesState extends State<CommunityChallenges> {
                       ),
                       onTap: () {
                         Navigator.of(context)
-                            .push(MaterialPageRoute<OverlayPopData>(
+                            .push(MaterialPageRoute<OverlayResult>(
                                 builder: (context) => Challenge(
                                       challenge: challenges[i],
                                       hasNext: false,
                                     )))
-                            .then((OverlayPopData popData) {});
+                            .then((OverlayResult popData) {});
                       },
                     )),
           ),
