@@ -76,7 +76,7 @@ class _EditorPlacerState extends State<EditorPlacer> {
 
   Widget _dragTileBuilder(BuildContext context, List<EditorTool> candidateData,
       List rejectedData, int x, int y) {
-    if (candidateData.length == 0) return const SizedBox.expand();
+    if (candidateData.isEmpty) return const SizedBox.expand();
 
     return _toolView(candidateData[0]);
   }
@@ -157,7 +157,7 @@ class _EditorPlacerState extends State<EditorPlacer> {
                 child: AspectRatio(
                     aspectRatio: 12.0 / 9.0,
                     child: InputGridOverlay<EditorTool>(
-                      child: ValueListenableBuilder<Game>(
+                      child: ValueListenableBuilder<GameState>(
                           valueListenable:
                               widget.editorGameController.gameStream,
                           builder: (context, value, child) {
