@@ -1,8 +1,11 @@
+import 'dart:async';
+
 import 'package:flutter/material.dart';
+import 'package:flutter/scheduler.dart';
 import 'package:flutter/services.dart';
 import 'package:nyanya_rocket/models/multiplayer_board.dart';
 import 'package:nyanya_rocket/screens/multiplayer/device_multiplayer_game_controller.dart';
-import 'package:nyanya_rocket/screens/multiplayer/widgets/event_roulette.dart';
+import 'package:nyanya_rocket/screens/multiplayer/game_widgets/event_wheel.dart';
 import 'package:nyanya_rocket/widgets/arrow_image.dart';
 import 'package:nyanya_rocket/widgets/countdown.dart';
 import 'package:nyanya_rocket/widgets/game_view/animated_game_view.dart';
@@ -10,24 +13,22 @@ import 'package:nyanya_rocket/widgets/input_grid_overlay.dart';
 import 'package:nyanya_rocket/widgets/score_box.dart';
 import 'package:nyanya_rocket_base/nyanya_rocket_base.dart';
 
-class DeviceMultiplayer extends StatefulWidget {
+class LocalDuel extends StatefulWidget {
   final MultiplayerBoard board;
   final List<String> players;
   final Duration duration;
 
-  DeviceMultiplayer(
+  LocalDuel(
       {@required this.board, @required this.players, @required this.duration});
 
   @override
-  _DeviceMultiplayerState createState() => _DeviceMultiplayerState();
+  _LocalDuelState createState() => _LocalDuelState();
 }
 
-class _DeviceMultiplayerState extends State<DeviceMultiplayer>
-    with SingleTickerProviderStateMixin {
+class _LocalDuelState extends State<LocalDuel> {
   LocalMultiplayerGameController _localMultiplayerController;
   bool _displayRoulette = false;
-  GameEvent _rouletteEvent = GameEvent.None;
-  AnimationController _animationController;
+  FixedExtentScrollController _scrollController = FixedExtentScrollController();
 
   @override
   void initState() {
@@ -36,29 +37,24 @@ class _DeviceMultiplayerState extends State<DeviceMultiplayer>
     _localMultiplayerController = LocalMultiplayerGameController(
         board: widget.board, onGameEvent: _handleGameEvent);
 
-    _animationController = AnimationController(
-        duration: const Duration(milliseconds: 1500), vsync: this)
-      ..addStatusListener((AnimationStatus status) {
-        if (status == AnimationStatus.completed) {
-          setState(() {
-            _displayRoulette = false;
-          });
-        }
-      });
-
     SystemChrome.setPreferredOrientations([
       DeviceOrientation.landscapeLeft,
       DeviceOrientation.landscapeRight,
     ]).catchError((Object error) {});
+
+    SystemChrome.setEnabledSystemUIOverlays([]);
   }
 
   @override
   void dispose() {
-    super.dispose();
-
-    _localMultiplayerController.close();
+    _localMultiplayerController.dispose();
 
     SystemChrome.setPreferredOrientations([]).catchError((Object error) {});
+
+    SystemChrome.setEnabledSystemUIOverlays(
+        [SystemUiOverlay.top, SystemUiOverlay.bottom]);
+
+    super.dispose();
   }
 
   void _handleDrop(int x, int y, Tile tile) {
@@ -71,7 +67,7 @@ class _DeviceMultiplayerState extends State<DeviceMultiplayer>
 
   Widget _dragTileBuilder(BuildContext context, List<Arrow> candidateData,
       List rejectedData, int x, int y) {
-    if (candidateData.length == 0) return const SizedBox.expand();
+    if (candidateData.isEmpty) return const SizedBox.expand();
 
     return ArrowImage(
       player: candidateData[0].player,
@@ -80,31 +76,60 @@ class _DeviceMultiplayerState extends State<DeviceMultiplayer>
     );
   }
 
-  Widget _draggableArrow(PlayerColor player, Direction direction) =>
-      Draggable<Arrow>(
+  Widget _draggableArrow(PlayerColor player, Direction direction) {
+    return Material(
+      elevation: 8.0,
+      child: Draggable<Arrow>(
           maxSimultaneousDrags: 1,
           feedback: const SizedBox.shrink(),
           child: ArrowImage(
             player: player,
             direction: direction,
           ),
-          data: Arrow(player: player, direction: direction));
+          data: Arrow(player: player, direction: direction)),
+    );
+  }
 
   void _handleGameEvent(GameEvent event) {
-    _animationController.forward(from: 0);
-    _rouletteEvent = event;
-    setState(() {
-      _displayRoulette = true;
-    });
+    // Not 0 to have a card above and under the starting position on the wheel.
+    _scrollController.jumpToItem(1);
+
+    if (event != GameEvent.None) {
+      setState(() {
+        _displayRoulette = true;
+      });
+
+      // The scroll controller won't animate if the wheel hasn't been attached
+      // before.
+      // Using a Wheel as a state attribute doesn't work because its build
+      // function creates a new ScrollView every time.
+      // Possible fixes:
+      // - make Wheel inherit ScrollView directly.
+      // - make Wheel keep its ScrollView across redraws.
+      SchedulerBinding.instance.addPostFrameCallback((_) {
+        _scrollController.animateToItem(
+            // Not * 4 to have a card above and under on the wheel.
+            (GameEvent.values.length - 1) * 3 + event.index - 1,
+            duration: Duration(milliseconds: 1500),
+            curve: Curves.decelerate);
+      });
+
+      Timer(Duration(seconds: 3), () {
+        setState(() {
+          _displayRoulette = false;
+        });
+      });
+    }
   }
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      resizeToAvoidBottomPadding: false,
-      body: Stack(
+    return Container(
+      color: Colors.white,
+      child: Stack(
         children: <Widget>[
           Row(
+            mainAxisAlignment: MainAxisAlignment.spaceEvenly,
             children: <Widget>[
               Expanded(
                 child: Padding(
@@ -120,14 +145,12 @@ class _DeviceMultiplayerState extends State<DeviceMultiplayer>
                               _draggableArrow(PlayerColor.Blue, Direction.Up)),
                       Expanded(
                           child: Container(
-                              child: StreamBuilder<int>(
-                                  stream: _localMultiplayerController
-                                      .scoreStreams[PlayerColor.Blue.index]
-                                      .stream,
-                                  initialData: 0,
-                                  builder: (context, snapshot) {
+                              child: ValueListenableBuilder<int>(
+                                  valueListenable: _localMultiplayerController
+                                      .scoreStreams[PlayerColor.Blue.index],
+                                  builder: (context, score, snapshot) {
                                     return ScoreBox(
-                                      score: snapshot.data,
+                                      score: score,
                                       label: widget.players[0],
                                       color: Colors.blue,
                                     );
@@ -146,28 +169,36 @@ class _DeviceMultiplayerState extends State<DeviceMultiplayer>
                   flex: 4,
                   child: Column(
                     children: <Widget>[
-                      StreamBuilder<Object>(
-                          stream: _localMultiplayerController.timeStream.stream,
-                          initialData: Duration.zero,
-                          builder: (context, snapshot) {
+                      ValueListenableBuilder<Duration>(
+                          valueListenable:
+                              _localMultiplayerController.timeStream,
+                          builder: (context, remaining, snapshot) {
                             return Padding(
                               padding:
                                   const EdgeInsets.symmetric(vertical: 8.0),
                               child: Countdown(
-                                remaining: snapshot.data,
+                                remaining: remaining,
                               ),
                             );
                           }),
                       Flexible(
-                        child: AspectRatio(
-                            aspectRatio: 12.0 / 9.0,
-                            child: InputGridOverlay<Arrow>(
-                              child: AnimatedGameView(
-                                game: _localMultiplayerController.gameStream,
-                              ),
-                              onDrop: _handleDrop,
-                              previewBuilder: _dragTileBuilder,
-                            )),
+                        child: Padding(
+                          padding:
+                              const EdgeInsets.fromLTRB(0.0, 0.0, 0.0, 8.0),
+                          child: Material(
+                            elevation: 8.0,
+                            child: AspectRatio(
+                                aspectRatio: 12.0 / 9.0,
+                                child: InputGridOverlay<Arrow>(
+                                  child: AnimatedGameView(
+                                    game:
+                                        _localMultiplayerController.gameStream,
+                                  ),
+                                  onDrop: _handleDrop,
+                                  previewBuilder: _dragTileBuilder,
+                                )),
+                          ),
+                        ),
                       ),
                     ],
                   )),
@@ -185,14 +216,12 @@ class _DeviceMultiplayerState extends State<DeviceMultiplayer>
                               _draggableArrow(PlayerColor.Red, Direction.Up)),
                       Expanded(
                           child: Container(
-                              child: StreamBuilder<int>(
-                                  stream: _localMultiplayerController
-                                      .scoreStreams[PlayerColor.Red.index]
-                                      .stream,
-                                  initialData: 0,
-                                  builder: (context, snapshot) {
+                              child: ValueListenableBuilder<int>(
+                                  valueListenable: _localMultiplayerController
+                                      .scoreStreams[PlayerColor.Red.index],
+                                  builder: (context, score, snapshot) {
                                     return ScoreBox(
-                                      score: snapshot.data,
+                                      score: score,
                                       label: widget.players[1],
                                       color: Colors.red,
                                     );
@@ -213,10 +242,11 @@ class _DeviceMultiplayerState extends State<DeviceMultiplayer>
             visible: _displayRoulette,
             child: Center(
                 child: Container(
+              width: 450,
               height: 200,
-              child: EventRoulette(
-                  animationController: _animationController,
-                  finalEvent: _rouletteEvent),
+              child: EventWheel(
+                scrollController: _scrollController,
+              ),
             )),
           )
         ],
