@@ -1,3 +1,5 @@
+import 'package:firebase_auth/firebase_auth.dart' as auth;
+import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:nyanya_rocket/localization/nyanya_localizations.dart';
 import 'package:nyanya_rocket/models/user.dart';
@@ -137,44 +139,134 @@ class AccountManagement extends StatelessWidget {
       ),
       body: Consumer<User>(
         builder: (innerContext, user, _) => ListView(
-              children: <Widget>[
-                ListTile(
-                  title: Text(
-                      '${NyaNyaLocalizations.of(context).loginStatusLabel}: ${user.isConnected ? NyaNyaLocalizations.of(context).connectedStatusLabel : NyaNyaLocalizations.of(context).disconnectedStatusLabel}'),
-                  subtitle: Text(user.isConnected
-                      ? NyaNyaLocalizations.of(context).signOutLabel
-                      : NyaNyaLocalizations.of(context).signInLabel),
-                  onTap: () {
-                    if (user.isConnected) {
-                      _showConfirmDialog(
-                              context,
-                              NyaNyaLocalizations.of(context)
-                                  .signOutDialogTitle,
-                              Text(NyaNyaLocalizations.of(context)
-                                  .signOutDialogText))
-                          .then((bool confirmed) {
-                        if (confirmed != null && confirmed) {
-                          user.signOut();
-                        }
-                      });
-                    } else {
-                      Navigator.push<bool>(
-                              context,
-                              MaterialPageRoute(
-                                  builder: (BuildContext context) =>
-                                      PrivacyPolicyPrompt(askUser: true)))
-                          .then((bool confirmed) {
-                        if (confirmed != null && confirmed) {
-                          user.signInAnonymously();
-                        }
-                      });
+          children: <Widget>[
+            ListTile(
+              title: Text(
+                  '${NyaNyaLocalizations.of(context).loginStatusLabel}: ${user.isConnected ? NyaNyaLocalizations.of(context).connectedStatusLabel : NyaNyaLocalizations.of(context).disconnectedStatusLabel}'),
+              subtitle: Text(user.isConnected
+                  ? NyaNyaLocalizations.of(context).signOutLabel
+                  : NyaNyaLocalizations.of(context).signInLabel),
+              onTap: () {
+                if (user.isConnected) {
+                  _showConfirmDialog(
+                          context,
+                          NyaNyaLocalizations.of(context).signOutDialogTitle,
+                          Text(NyaNyaLocalizations.of(context)
+                              .signOutDialogText))
+                      .then((bool confirmed) {
+                    if (confirmed ?? false) {
+                      user.signOut();
                     }
-                  },
-                ),
-                _buildDisplayNameTile(innerContext, user)
-              ],
+                  });
+                } else {
+                  promptSignUp(context);
+                }
+              },
             ),
+            _buildDisplayNameTile(innerContext, user),
+            ListTile(
+              title: Text(NyaNyaLocalizations.of(context).privacyPolicyLabel),
+              onTap: () {
+                Navigator.push<bool>(
+                    context,
+                    MaterialPageRoute(
+                        builder: (BuildContext context) =>
+                            PrivacyPolicyPrompt(askUser: false)));
+              },
+            ),
+          ],
+        ),
       ),
     );
+  }
+
+  static Future<bool> promptSignUp(BuildContext context) {
+    String displayName;
+    final _formKey = GlobalKey<FormState>();
+
+    return showDialog<bool>(
+        context: context,
+        builder: (BuildContext context) {
+          return AlertDialog(
+            contentPadding: const EdgeInsets.all(16.0),
+            title: Text(NyaNyaLocalizations.of(context).displayNameDialogTitle),
+            content: Column(
+              mainAxisSize: MainAxisSize.min,
+              children: <Widget>[
+                Form(
+                  key: _formKey,
+                  child: TextFormField(
+                    autofocus: true,
+                    autovalidate: true,
+                    maxLength: 24,
+                    validator: (String value) {
+                      if (!AccountManagement.displayNameRegExp
+                          .hasMatch(value)) {
+                        return NyaNyaLocalizations.of(context)
+                            .displayNameFormatText;
+                      }
+
+                      return null;
+                    },
+                    onSaved: (String value) {
+                      displayName = value;
+                    },
+                  ),
+                ),
+                const SizedBox(height: 8.0),
+                RichText(
+                  textAlign: TextAlign.center,
+                  text: TextSpan(
+                    style: Theme.of(context).textTheme.bodyText1,
+                    children: <TextSpan>[
+                      TextSpan(
+                          text: NyaNyaLocalizations.of(context)
+                              .privacyPolicySignUpText),
+                      TextSpan(
+                          text: NyaNyaLocalizations.of(context)
+                              .privacyPolicyLabel,
+                          style: TextStyle(
+                              color: Colors.blue,
+                              decoration: TextDecoration.underline),
+                          recognizer: TapGestureRecognizer()
+                            ..onTap = () {
+                              Navigator.push(
+                                  context,
+                                  MaterialPageRoute(
+                                      builder: (BuildContext context) =>
+                                          const PrivacyPolicyPrompt(
+                                              askUser: false)));
+                            })
+                    ],
+                  ),
+                )
+              ],
+            ),
+            actions: <Widget>[
+              FlatButton(
+                  child: Text(NyaNyaLocalizations.of(context)
+                      .confirmLabel
+                      .toUpperCase()),
+                  onPressed: () {
+                    if (_formKey.currentState.validate()) {
+                      _formKey.currentState.save();
+                      Navigator.pop(context, true);
+
+                      auth.FirebaseAuth.instance
+                          .signInAnonymously()
+                          .then((user) {
+                        user.user.updateProfile(displayName: displayName);
+                      });
+                    }
+                  }),
+              FlatButton(
+                  child: Text(
+                      NyaNyaLocalizations.of(context).cancel.toUpperCase()),
+                  onPressed: () {
+                    Navigator.pop(context, false);
+                  }),
+            ],
+          );
+        });
   }
 }
