@@ -6,18 +6,13 @@ import 'package:nyanya_rocket/screens/settings/dark_mode.dart';
 import 'package:nyanya_rocket/screens/settings/language.dart';
 import 'package:provider/provider.dart';
 
-class Settings extends StatefulWidget {
-  @override
-  _SettingsState createState() {
-    return _SettingsState();
-  }
-}
+import 'region.dart';
 
-class _SettingsState extends State<Settings> {
-  @override
-  void initState() {
-    super.initState();
-  }
+class Settings extends StatelessWidget {
+  static const Map<Regions, String> _regionLabels = {
+    Regions.usEast: 'US East',
+    Regions.euWest: 'Europe West'
+  };
 
   @override
   Widget build(BuildContext context) {
@@ -30,9 +25,7 @@ class _SettingsState extends State<Settings> {
             SwitchListTile(
               title: Text(NyaNyaLocalizations.of(context).darkModeLabel),
               onChanged: (bool value) {
-                setState(() {
-                  Provider.of<DarkMode>(context).enabled = value;
-                });
+                Provider.of<DarkMode>(context, listen: false).enabled = value;
               },
               value: Provider.of<DarkMode>(context).enabled,
             ),
@@ -59,22 +52,46 @@ class _SettingsState extends State<Settings> {
                     ),
                   ],
                   onChanged: (String language) {
-                    Provider.of<Language>(context).value = language;
+                    Provider.of<Language>(context, listen: false).value =
+                        language;
+                  }),
+            ),
+            ListTile(
+              title: Text(NyaNyaLocalizations.of(context).regionLabel),
+              trailing: DropdownButton<Regions>(
+                  value: Provider.of<Region>(context).value,
+                  items: <DropdownMenuItem<Regions>>[
+                    DropdownMenuItem(
+                      child: Text(
+                          'Auto (${_regionLabels[Region.automaticValue()]})'),
+                      value: Regions.auto,
+                    ),
+                    DropdownMenuItem(
+                      child: Text(_regionLabels[Regions.euWest]),
+                      value: Regions.euWest,
+                    ),
+                    DropdownMenuItem(
+                      child: Text(_regionLabels[Regions.usEast]),
+                      value: Regions.usEast,
+                    ),
+                  ],
+                  onChanged: (Regions region) {
+                    Provider.of<Region>(context, listen: false).value = region;
                   }),
             ),
             Consumer<User>(
               builder: (context, user, _) => ListTile(
-                    title: Text(
-                        NyaNyaLocalizations.of(context).accountManagementLabel),
-                    subtitle: Text(
-                        '${NyaNyaLocalizations.of(context).loginStatusLabel}: ${user.isConnected ? NyaNyaLocalizations.of(context).connectedStatusLabel : NyaNyaLocalizations.of(context).disconnectedStatusLabel}'),
-                    onTap: () {
-                      Navigator.push(context,
-                          MaterialPageRoute(builder: (BuildContext context) {
-                        return AccountManagement();
-                      }));
-                    },
-                  ),
+                title: Text(
+                    NyaNyaLocalizations.of(context).accountManagementLabel),
+                subtitle: Text(
+                    '${NyaNyaLocalizations.of(context).loginStatusLabel}: ${user.isConnected ? NyaNyaLocalizations.of(context).connectedStatusLabel : NyaNyaLocalizations.of(context).disconnectedStatusLabel}'),
+                onTap: () {
+                  Navigator.push(context,
+                      MaterialPageRoute(builder: (BuildContext context) {
+                    return AccountManagement();
+                  }));
+                },
+              ),
             ),
           ],
         ));
