@@ -5,7 +5,7 @@ class Puzzle extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return ListView(
-      padding: const EdgeInsets.all(8),
+      padding: const EdgeInsets.all(8.0),
       children: <Widget>[
         Text(NyaNyaLocalizations.of(context).puzzleTutorialText),
         Card(
