@@ -16,26 +16,21 @@ class NotImplemented extends StatelessWidget {
           NyaNyaLocalizations.of(context).notImplementedText,
           textAlign: TextAlign.center,
         ),
-        Row(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            IconButton(
-              icon: Icon(
-                Icons.thumb_up,
-                color: Colors.green,
-              ),
-              onPressed: () {
-                final DocumentReference postRef =
-                    Firestore.instance.document('feature_requests/$featureId');
-                postRef.updateData({'thumbs_up': FieldValue.increment(1)});
-              },
-            ),
-          ],
+        IconButton(
+          icon: Icon(
+            Icons.thumb_up,
+            color: Colors.green,
+          ),
+          onPressed: () {
+            final DocumentReference postRef =
+                FirebaseFirestore.instance.doc('feature_requests/$featureId');
+            postRef.update({'thumbs_up': FieldValue.increment(1)});
+          },
         ),
         StreamBuilder<DocumentSnapshot>(
-          stream: Firestore.instance
+          stream: FirebaseFirestore.instance
               .collection('feature_requests')
-              .document(featureId)
+              .doc(featureId)
               .snapshots(),
           builder:
               (BuildContext context, AsyncSnapshot<DocumentSnapshot> snapshot) {
@@ -44,9 +39,10 @@ class NotImplemented extends StatelessWidget {
             }
 
             switch (snapshot.connectionState) {
+              case ConnectionState.active:
               case ConnectionState.done:
                 return Text(
-                  snapshot.data['thumbs_up'].toString(),
+                  snapshot.data.get('thumbs_up').toString(),
                   textAlign: TextAlign.center,
                 );
                 break;
