@@ -11,37 +11,38 @@ class ScoreBox extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Card(
+      elevation: 8.0,
       color: color,
       child: Column(
+        mainAxisSize: MainAxisSize.min,
         children: <Widget>[
-          Flexible(
-            flex: 1,
-            child: Padding(
-              padding: const EdgeInsets.fromLTRB(8, 0, 8, 0),
-              child: Container(
-                  child: Center(child: Text(label ?? '')),
-                  constraints: BoxConstraints.expand(),
-                  decoration: BoxDecoration(
-                      color: Colors.white,
-                      borderRadius: BorderRadius.all(Radius.circular(8.0)))),
-            ),
-          ),
-          Flexible(
-            flex: 3,
-            child: Padding(
-              padding: const EdgeInsets.fromLTRB(8.0, 0, 8.0, 8.0),
-              child: Container(
-                child: Center(
-                  child: Text(
-                    score.toString(),
-                    style: TextStyle(color: Colors.white, fontSize: 30),
-                  ),
+          Padding(
+            padding: const EdgeInsets.symmetric(horizontal: 4.0),
+            child: Container(
+                child: Text(
+                  label ?? '',
+                  textAlign: TextAlign.center,
                 ),
-                constraints: BoxConstraints.expand(),
+                padding: const EdgeInsets.all(2.0),
+                constraints: const BoxConstraints(minWidth: 80),
                 decoration: BoxDecoration(
-                    color: Colors.black,
-                    borderRadius: BorderRadius.all(Radius.circular(8.0))),
+                    color: Colors.white,
+                    borderRadius: BorderRadius.all(Radius.circular(4.0)))),
+          ),
+          Padding(
+            padding: const EdgeInsets.fromLTRB(4.0, 0, 4.0, 4.0),
+            child: Container(
+              child: Center(
+                child: Text(
+                  score.toString(),
+                  style: TextStyle(color: Colors.white, fontSize: 30),
+                ),
               ),
+              padding: const EdgeInsets.all(8.0),
+              constraints: const BoxConstraints(minWidth: 80),
+              decoration: BoxDecoration(
+                  color: Colors.black,
+                  borderRadius: BorderRadius.all(Radius.circular(4.0))),
             ),
           )
         ],
